+++ lib/diff.mli
@@ -7,18 +7,18 @@ type value = {
     (Types.value_description, Types.value_description atomic_modification) t;
 }
 
-type class_modification = Unsupported
-
 and class_ = {
   cname : string;
-  cdiff : (Types.class_declaration, class_modification) t;
+  cdiff :
+    (Types.class_declaration, Types.class_declaration atomic_modification) t;
 }
-
-type class_type_modification = Unsupported
 
 and cltype = {
   ctname : string;
-  ctdiff : (Types.class_type_declaration, class_type_modification) t;
+  ctdiff :
+    ( Types.class_type_declaration,
+      Types.class_type_declaration atomic_modification )
+    t;
 }
 
 type type_modification =